pubsub is now shared logic, and google.ts is google specific
